backend/requirements.txt
# Add FastAPI, psycopg2, and other dependencies here
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
async-property==0.2.2
asyncpg==0.30.0
cached-property==1.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
deprecation==2.1.0
ecdsa==0.19.1
fastapi==0.115.13
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
idna==3.10
jwcrypto==1.5.6
keycloak==3.1.5
packaging==25.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-jose==3.5.0
python-keycloak==5.5.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
